refactor(polytope_playground): extract color helpers into color module
Moves hsv_to_rgb, the per-edge palette, and w_depth_color (plus their tests) out of main.rs into color.rs, leaning the comments in the process. compute_cell_strengths stays in main.rs for now (it goes with the geometry helpers).
